// src/app/models/website.ts
export class Website {
  website_id: string;
  code: string;
  name: string;
  default_group_id: string;
  sort_order: number;
  is_default: number;
  store_group_name: string;
  store_view_name: string;
  //  <id>1</id>
  // <code>base</code>
  // <name>Main Website</name>
  // <default_group_id>1</default_group_id>
}